scripts: many improvements
- Standardize environment loading
- Accept arguments
  all arguments are passed through call stack
  each script picks the ones it understands
- Add arguments:
  - `--no-rebuild`: do not rebuild containers
    Implies no container removal when used with pristine.sh
  - `--no-python-purge`: do not remove Python venv
- Various fixes, mostly wrong paths

+++ scripts/data_purge.sh
@@ -3,6 +3,10 @@
 set -e #ux
 shopt -s dotglob
 
+## Record directory we are called from
+cwd="$(pwd)"
+## Change to project directory
+cd "$(dirname "$0")/.."
 source .env
 
 echo "-- Delete data"
@@ -10,5 +14,10 @@ sudo rm -rf "$LOG_DATA_DIR"/*
 sudo rm -rf "$DBA_DATA_DIR"/*
 sudo rm -rf "$APP_DATA_DIR"/*
 sudo find data/ssl/ -type f | sudo xargs rm -vf
-sudo find data/ssl/letsencrypt/accounts/acme-v02.api.letsencrypt.org \
-        - type d | sudo xargs rmdir -v
+if [[ -d "data/ssl/letsencrypt/accounts/acme-v02.api.letsencrypt.org" ]]; then
+  sudo find data/ssl/letsencrypt/accounts/acme-v02.api.letsencrypt.org/ \
+    -type d | sudo xargs rmdir -v
+fi
+
+## and go back
+cd "$cwd"
